requirements.txt
# Web frameworks
django==4.2.7
fastapi==0.104.1
uvicorn==0.23.2

# Database and vector storage
faiss-cpu==1.7.4
chromadb==0.4.18
sqlalchemy==2.0.23
sqlitedict==2.1.0

# ML and embeddings
torch==2.1.0
torchvision==0.16.0
clip-by-openai==0.1.1.5
transformers==4.35.0
pillow==10.1.0
numpy==1.24.4
imagebind-package==0.1.1

# Utilities
python-multipart==0.0.6
pydantic==2.4.2
psutil==5.9.6
loguru==0.7.2
tqdm==4.66.1
matplotlib==3.8.1
pandas==2.1.2
